requirements.txt
flask==3.1.1
openai==1.55.3
python-dotenv==1.0.1
python-decouple==3.8
gunicorn==21.2.0
